requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==2.1
django-clickhouse==1.0.0
django-pg-bulk-update==3.2.4
django-pg-returning==1.2.4
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
infi.clickhouse-orm==1.3.0
iso8601==0.1.12
kombu==4.6.8
lxml==4.5.0
opcua==0.98.9
psycopg2-binary==2.8.4
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
statsd==3.3.0
typing==3.7.4.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
